+++ inst/doc/Introduction.R
@@ -8,11 +8,11 @@ R <- function() knitr::include_graphics("Rlogo.png", dpi = 5000)
 ## ----echo=FALSE, out.width = '33%'--------------------------------------------
 knitr::include_graphics("hex-hce.png")
 
-## ----eval=TRUE----------------------------------------------------------------
+## ----eval = TRUE--------------------------------------------------------------
 library(hce)
 packageVersion("hce")
 
-## ----eval=FALSE, include=FALSE------------------------------------------------
+## ----eval = FALSE, include = FALSE--------------------------------------------
 # devtools::load_all()
 
 ## -----------------------------------------------------------------------------
@@ -32,22 +32,24 @@ class(dat)
 head(dat)
 
 ## -----------------------------------------------------------------------------
-unique(dat[, c("GROUP", "GROUPN")])
+unique(dat[, c("GROUP", "PARAMN")])
 
 ## -----------------------------------------------------------------------------
 set.seed(2022)
 n <- 100
-dat <- hce(GROUP = rep(x = c("I", "II", "III"), each = 100), 
-           TRTP = sample(x = c("A", "P"), size = n*3, replace = TRUE), ORD = c("III", "II", "I"))
-unique(dat[, c("GROUP", "GROUPN")])
+GROUP = rep(x = c("I", "II", "III"), each = 100) 
+GROUP <- factor(GROUP, levels = c("III", "II", "I"))
+dat <- hce(GROUP = GROUP, 
+           TRTP = sample(x = c("Active", "Control"), size = n*3, replace = TRUE))
+unique(dat[, c("GROUP", "PARAMN")])
 
 ## -----------------------------------------------------------------------------
 data(HCE1)
 unique(HCE1$GROUP)
 
 ## -----------------------------------------------------------------------------
-HCE <- hce(GROUP = HCE1$GROUP, TRTP = HCE1$TRTP, AVAL0 = HCE1$AVAL0, 
-           ORD = c("TTE1", "TTE2", "TTE3", "TTE4", "C"))
+HCE <- hce(GROUP = factor(HCE1$GROUP, levels = c("TTE1", "TTE2", "TTE3", "TTE4", "C")), 
+           TRTP = HCE1$TRTP, AVAL0 = HCE1$AVAL0, PADY = 1080)
 class(HCE)
 head(HCE)
 
@@ -80,12 +82,17 @@ head(dat3)
 methods(class = "hce")
 
 ## -----------------------------------------------------------------------------
-HCE <- hce(GROUP = HCE3$GROUP, TRTP = HCE3$TRTP,
-           ORD = c("TTE1", "TTE2", "TTE3", "TTE4", "C"))
+HCE <- hce(GROUP = factor(HCE3$GROUP, levels = c("TTE1", "TTE2", "TTE3", "TTE4", "C")), 
+           TRTP = HCE3$TRTP, AVAL0 = HCE3$AVAL0, PADY = 1080)
 calcWO(HCE)
-summaryWO(HCE)
 calcWINS(HCE)  
 HCE$TRTP <- factor(HCE$TRTP, levels = c("P", "A"))
 plot(HCE, fill = TRUE, col = "#865A4F", type = 'l', lwd = 2)
 abline(a = 0, b = 1, lwd = 2, col = "#999999", lty = 2)
+
+## -----------------------------------------------------------------------------
+methods(class = "adhce")
+
+## -----------------------------------------------------------------------------
+summaryWO(HCE)